simplify error handler (#22)
# Pull Request Template

## Description

Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. List any
dependencies that are required for this change.

## Type of change:

Please select the appropriate tags in the PR if necessary.

# Motivation

Keeping thing up-to-date

# How has this been tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce, and the Operative System used.

## Checklist:

- [ ] My code follows the style guidelines of this project.
- [ ] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] Select the appropriate tags in the PR if necessary.